% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_JS.R
\name{scale_fill_JS}
\alias{scale_fill_JS}
\title{JS scale fill for ggplot2}
\usage{
scale_fill_JS(
  theme = "JS",
  tech_key = list(JS = c("#BDB76B", "#556B2F", "#F4A460", "#FFFACD", "#7FFFD4"), Bright
    = c("#FF0000", "#00FF00", "#FFFF00", "#00FFFF", "#FF00FF", "#FF00FF"))
)
}
\description{
JS scale fill, which is built on Ricardo Bion's code for ggtech
}
\examples{
scale_fill_JS(theme="JS")
}
