% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_JS.R
\name{theme_JS}
\alias{theme_JS}
\title{JS themes  for ggplot2}
\usage{
theme_JS(
  theme = "JS",
  tech_key = list(JS = list(family_title = "PicoBlackAl", family_text =
    "[z] Arista Light", colour_title = "#5380E4", colour_text = "#535353"))
)
}
\description{
JS theme, which is built on Ricardo Bion's code for ggtech
}
\examples{
theme_JS(theme="JS")
}
