% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_JS_fancy.R
\name{theme_JS_fancy}
\alias{theme_JS_fancy}
\title{JS theme for ggplot2}
\usage{
theme_JS_fancy(base_size = 2.5)
}
\arguments{
\item{base_size}{base_size}
}
\value{
A theme element
}
\description{
Theme JS fancy, which is built on Ricardo Bion's code for ggtech
}
